updated fixed some minor bugs
Today I noticed that I still had a couple of tasks calling the buttons sub-routine that used to provide a 2nd window with radio buttons for making a selection. With IE8 and higher it doesn't work. I re-wrote that on the the sub routines in which I found the problem.